minimal testing proxy selector
