+++ apm-agent-core/src/test/java/co/elastic/apm/agent/util/PrivilegedActionUtilsTest.java
@@ -18,7 +18,6 @@
  */
 package co.elastic.apm.agent.util;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledForJreRange;
 import org.junit.jupiter.api.condition.JRE;
@@ -30,9 +29,11 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.security.AccessController;
+import java.security.Permission;
 import java.security.PrivilegedAction;
 import java.security.ProtectionDomain;
 import java.util.Map;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 import static co.elastic.apm.agent.testutils.assertions.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -40,12 +41,18 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 @EnabledForJreRange(max = JRE.JAVA_17, disabledReason = "SecurityManager is not supported anymore")
 class PrivilegedActionUtilsTest {
 
-    @BeforeEach
-    void beforeAll() {
-        // those tests rely on the SM to be disabled on startup
-        TestSecurityManager.disable();
+    private static final AtomicBoolean enabled = new AtomicBoolean(false);
+
+    private static void enableSecurityManager() {
+        enabled.set(false);
+        System.setSecurityManager(new TestSecurityManager(enabled));
+        enabled.set(true);
     }
 
+    private static void disableSecurityManager() {
+        enabled.set(false);
+        System.setSecurityManager(null);
+    }
 
     @Test
     void getEnv() {
@@ -53,7 +60,7 @@ class PrivilegedActionUtilsTest {
         String envKey = envMap.keySet().stream().findFirst().get();
         String envValue = envMap.get(envKey);
 
-        testWithAndWithoutSecurityManager(()->{
+        testWithAndWithoutSecurityManager(() -> {
             assertThat(PrivilegedActionUtils.getEnv(envKey)).isEqualTo(envValue);
             assertThat(PrivilegedActionUtils.getEnv()).containsAllEntriesOf(envMap);
         });
@@ -63,13 +70,13 @@ class PrivilegedActionUtilsTest {
     @Test
     void getClassLoader() {
         ClassLoader cl = PrivilegedActionUtilsTest.class.getClassLoader();
-        testWithAndWithoutSecurityManager(()-> assertThat(PrivilegedActionUtils.getClassLoader(PrivilegedActionUtilsTest.class)).isSameAs(cl));
+        testWithAndWithoutSecurityManager(() -> assertThat(PrivilegedActionUtils.getClassLoader(PrivilegedActionUtilsTest.class)).isSameAs(cl));
     }
 
     @Test
     void getProtectionDomain() {
         ProtectionDomain pd = PrivilegedActionUtilsTest.class.getProtectionDomain();
-        testWithAndWithoutSecurityManager(()-> assertThat(PrivilegedActionUtils.getProtectionDomain(PrivilegedActionUtilsTest.class)).isSameAs(pd));
+        testWithAndWithoutSecurityManager(() -> assertThat(PrivilegedActionUtils.getProtectionDomain(PrivilegedActionUtilsTest.class)).isSameAs(pd));
     }
 
     @Test
@@ -154,7 +161,12 @@ class PrivilegedActionUtilsTest {
         });
     }
 
-    private static void testPrivileged(Runnable task){
+    @Test
+    void getProxySelector() {
+        testWithAndWithoutSecurityManager(PrivilegedActionUtils::getDefaultProxySelector);
+    }
+
+    private static void testPrivileged(Runnable task) {
         AccessController.doPrivileged(new PrivilegedAction<Object>() {
             @Override
             public Object run() {
@@ -164,15 +176,49 @@ class PrivilegedActionUtilsTest {
         });
     }
 
-    void testWithAndWithoutSecurityManager(Runnable assertions){
+    void testWithAndWithoutSecurityManager(Runnable assertions) {
         assertions.run();
         try {
-            TestSecurityManager.enable();
+            enableSecurityManager();
             assertions.run();
         } finally {
-            TestSecurityManager.disable();
+            disableSecurityManager();
         }
         assertions.run();
     }
 
+    private static class TestSecurityManager extends SecurityManager {
+
+        private final AtomicBoolean enabled;
+
+        public TestSecurityManager(AtomicBoolean enabled) {
+            this.enabled = enabled;
+        }
+
+        @Override
+        public void checkPermission(Permission perm) {
+            if (!enabled.get()) {
+                return;
+            }
+            checkPrivileged();
+        }
+
+        @Override
+        public void checkPermission(Permission perm, Object context) {
+            if (!enabled.get()) {
+                return;
+            }
+            checkPrivileged();
+        }
+
+        private static void checkPrivileged() {
+            StackTraceElement[] stackTrace = new RuntimeException().getStackTrace();
+            for (StackTraceElement e : stackTrace) {
+                if (e.getClassName().equals("java.security.AccessController") && e.getMethodName().equals("doPrivileged")) {
+                    return;
+                }
+            }
+            throw new SecurityException("missing privileged action");
+        }
+    }
 }
